rename hooks2 to hooks
