+++ lexio/lib/components/RAGProvider/index.tsx
@@ -1,6 +1,6 @@
 import { RAGProvider } from './provider';
 import { useRAGStatus } from './hooks';
 import { RAGProvider2 } from './provider2';
-import { useLexio, useRAGMessages, useRAGSources } from './hooks2';
+import { useLexio, useRAGMessages, useRAGSources } from './hooks';
 
 export { RAGProvider, RAGProvider2, useRAGSources, useRAGStatus, useLexio, useRAGMessages };
